requirements.txt
# Core dependencies
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1

# Machine Learning and AI
spacy==3.7.2
ultralytics==8.0.196
easyocr==1.7.0
pytesseract==0.3.10

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
streamlit-folium==0.13.0

# Data Processing
pandas==2.1.3
plotly==5.17.0

# Geospatial
folium==0.14.0

# HTTP and API
requests==2.31.0
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# Optional dependencies for enhanced functionality
# Uncomment if you need additional features
# torch==2.1.0
# torchvision==0.16.0
# transformers==4.35.2
